refactor(auth): move library file for firebase
