+++ src/lib/firebase.js
@@ -11,7 +11,6 @@ const firebaseConfig = {
   appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID
 };
 
-console.log(firebaseConfig);
 if (getApps().length === 0) {
   initializeApp(firebaseConfig);
 }
